Initial split-up of DependsOn into fixture and method attributes
